requirements.txt
streamlit>=1.30.0
requests>=2.31.0
pandas>=2.0.0
openpyxl>=3.1.0
pdfplumber>=0.10.0
pytesseract>=0.3.10
Pillow>=10.0.0
jinja2>=3.1.0
reportlab>=4.0.0
python-dateutil>=2.8.0
